+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Added fh2 support
+- Added comments on CSV SBOM export output 
 ### Changed
 - Upgraded `scanoss.js` version to v0.26.0