+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added fh2 support
 - Added comments on CSV SBOM export output
 - Added option to copy file path to clipboard on cryptography search screen
+- Displayed detected cryptography for a selected file in cryptography search screen
 ### Changed
 - Upgraded `scanoss.js` version to v0.26.0